Refactor assets folder and create plantCard component
#10
