+++ src/assets/mocks/plantSize.json
