+++ src/assets/mocks/plants.json
