+++ src/components/AvatarModal/Avatars.ts
@@ -1,32 +1,32 @@
 export const Avatars = [
     {
       nickname: "Juninha",
-      img: require("../../images/Avatars/Juninha.png"),
+      img: require("../../assets/images/Avatars/Juninha.png"),
       path: "../../images/Avatars/Juninha.png",
     },
     {
       nickname: "Led",
-      img: require("../../images/Avatars/Led.png"),
+      img: require("../../assets/images/Avatars/Led.png"),
       path: "../../images/Avatars/Led.png",
     },
     {
       nickname: "Tito",
-      img: require("../../images/Avatars/Tito.png"),
+      img: require("../../assets/images/Avatars/Tito.png"),
       path: "../../images/Avatars/Tito.png",
     },
     {
       nickname: "Fernandinho",
-      img: require("../../images/Avatars/Fernandinho.png"),
+      img: require("../../assets/images/Avatars/Fernandinho.png"),
       path: "../../images/Avatars/Fernandinho.png",
     },
     {
       nickname: "Glaucia",
-      img: require("../../images/Avatars/Glaucia.png"),
+      img: require("../../assets/images/Avatars/Glaucia.png"),
       path: "../../images/Avatars/Glaucia.png",
     },
     {
       nickname: "Dengoso",
-      img: require("../../images/Avatars/Dengoso.png"),
+      img: require("../../assets/images/Avatars/Dengoso.png"),
       path: "../../images/Avatars/Dengoso.png",
     },
   ];
