+++ src/components/LeafButton/index.tsx
@@ -4,7 +4,7 @@ import { Colors } from "../../configs/colors";
 import { styles } from "./styles";
 
 const LeafButton = ({ isFocused = false }): React.ReactElement => {
-  const leaf = require("../../images/Leaf.png");
+  const leaf = require("../../assets/images/Leaf.png");
   const buttonStyle = isFocused
     ? { ...styles.leafButton, backgroundColor: Colors.green_android }
     : { ...styles.leafButton, backgroundColor: Colors.green_hunter };
