+++ src/components/PlantCard/index.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { View, Text } from "react-native";
+import { styles } from "./styles";
+
+const PlantCard = (): React.ReactElement => {
+  const search = require("../../assets/images/Search.png");
+
+  return (
+    <View style={styles.container}>
+      <Text>Plant</Text>
+    </View>
+  );
+};
+
+export default PlantCard;
